# core-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Make-Singer — AI-powered music creation platform.",
        "title": "Make-Singer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/master/final": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start an asynchronous final mastering job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master"
                ],
                "summary": "Start final mastering",
                "parameters": [
                    {
                        "description": "Master final request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MasterFinalRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/model.MasterFinalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/master/preview": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate a quick mastering preview of the mix",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master"
                ],
                "summary": "Master preview",
                "parameters": [
                    {
                        "description": "Master preview request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MasterPreviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MasterPreviewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/master/result/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the result of a completed mastering job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master"
                ],
                "summary": "Get master job result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MasterResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/master/status/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the current status and progress of a mastering job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master"
                ],
                "summary": "Get master job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MasterStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.JobStatus": {
            "type": "string",
            "enum": [
                "queued",
                "running",
                "succeeded",
                "failed",
                "canceled"
            ],
            "x-enum-varnames": [
                "JobStatusQueued",
                "JobStatusRunning",
                "JobStatusSucceeded",
                "JobStatusFailed",
                "JobStatusCanceled"
            ]
        },
...
        "model.MasterFinalRequest": {
            "type": "object",
            "required": [
                "mixSnapshot",
                "profile",
                "projectId",
                "stemUrls"
            ],
            "properties": {
                "mixSnapshot": {
                    "$ref": "#/definitions/model.MixSnapshot"
                },
                "profile": {
                    "enum": [
                        "clean",
                        "warm",
                        "loud"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MasterProfile"
                        }
                    ]
                },
                "projectId": {
                    "type": "string"
                },
                "stemUrls": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "vocalTakes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.VocalTake"
                    }
                }
            }
        },
        "model.MasterFinalResponse": {
            "type": "object",
            "properties": {
                "estimatedDuration": {
                    "type": "integer"
                },
                "jobId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                }
            }
        },
        "model.MasterPreviewRequest": {
            "type": "object",
            "required": [
                "mixSnapshot",
                "profile",
                "projectId",
                "stemUrls"
            ],
            "properties": {
                "mixSnapshot": {
                    "$ref": "#/definitions/model.MixSnapshot"
                },
                "previewStartTime": {
                    "type": "integer",
                    "minimum": 0
                },
                "profile": {
                    "enum": [
                        "clean",
                        "warm",
                        "loud"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MasterProfile"
                        }
                    ]
                },
                "projectId": {
                    "type": "string"
                },
                "stemUrls": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.MasterPreviewResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "expiresAt": {
                    "type": "string"
                },
                "fileUrl": {
                    "type": "string"
                }
            }
        },
        "model.MasterProfile": {
            "type": "string",
            "enum": [
                "clean",
                "warm",
                "loud"
            ],
            "x-enum-varnames": [
                "MasterProfileClean",
                "MasterProfileWarm",
                "MasterProfileLoud"
            ]
        },
        "model.MasterResultResponse": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "expiresAt": {
                    "type": "string"
                },
                "fileUrl": {
                    "type": "string"
                },
                "lufs": {
                    "type": "integer"
                },
                "peakDb": {
                    "type": "number"
                },
                "profile": {
                    "$ref": "#/definitions/model.MasterProfile"
                }
            }
        },
        "model.MasterStatusResponse": {
            "type": "object",
            "properties": {
                "currentStep": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                }
            }
        },
        "model.MixChannel": {
            "type": "object",
            "required": [
                "stemId"
            ],
            "properties": {
                "mute": {
                    "type": "boolean"
                },
                "solo": {
                    "type": "boolean"
                },
                "stemId": {
                    "type": "string"
                },
                "volumeDb": {
                    "type": "number",
                    "maximum": 12,
                    "minimum": -60
                }
            }
        },
        "model.MixPreset": {
            "type": "string",
            "enum": [
                "default",
                "vocal_friendly",
                "bass_heavy",
                "bright",
                "warm"
            ],
            "x-enum-varnames": [
                "MixPresetDefault",
                "MixPresetVocalFriendly",
                "MixPresetBassHeavy",
                "MixPresetBright",
                "MixPresetWarm"
            ]
        },
        "model.MixSnapshot": {
            "type": "object",
            "required": [
                "channels",
                "preset"
            ],
            "properties": {
                "channels": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.MixChannel"
                    }
                },
                "preset": {
                    "enum": [
                        "default",
                        "vocal_friendly",
                        "bass_heavy",
                        "bright",
                        "warm"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.MixPreset"
                        }
                    ]
                }
            }
        },
...
        "model.VocalTake": {
            "type": "object",
            "required": [
                "fileUrl",
                "sectionId",
                "takeId"
            ],
            "properties": {
                "fileUrl": {
                    "type": "string"
                },
                "offsetMs": {
                    "type": "integer"
                },
                "sectionId": {
                    "type": "string"
                },
                "takeId": {
                    "type": "string"
                }
            }
        },
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}